% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FramedClust.R
\name{FramedClust}
\alias{FramedClust}
\title{Framed Data Clustering}
\usage{
FramedClust(
  X,
  K,
  frame.width,
  first.frame = 1,
  last.frame = length(X) - frame.width + 1,
  method = c("linear.polylog", "kmeans", "Ckmeans.1d.dp")
)
}
\arguments{
\item{X}{The set of data points on which the search has to be conducted}

\item{K}{The Number of Clusters in each frame}

\item{frame.width}{The Number of Points from X to be included in each frame. The actual width of the frame may vary from iteration to iteration.}

\item{first.frame}{Starting index of the first frame to be clustered}

\item{last.frame}{Starting index of the last frame to be clustered}

\item{method}{the circular clustering method.
\code{"linear.polylog"}: fast and optimal, the default method;
  \code{"kmeans"}: based on heuristic k-means, fast but not necessarily optimal;
  \code{"Ckmeans.1d.dp"}: brute-force based on quadratic runtime algorithm from the Ckmeans.1d.dp. The algorithm is slow but optimal,
  included to provide a baseline.}
}
\value{
An object of class \code{"FramedClust"} which has a \code{plot}
 method. It is a list with the following components:

\item{cluster}{ A vector of clusters assigned to each element in x. Each cluster is indexed by an integer from 1 to k. NA represents the points from X that are not part of any cluster.}

\item{centers}{ A numeric vector of the  means for each cluster in the frame.}

\item{withinss}{	A numeric vector of the  within-cluster sum of squares for each cluster.}

\item{size}{	A vector of the  number of elements in each cluster.}

\item{totss}{	Total sum of  squared distances between each element and the sample mean. This statistic is not dependent on the clustering result.}

\item{tot.withinss}{	 Total sum of  within-cluster squared distances between each element and its cluster mean. This statistic is minimized given the number of clusters.}

\item{betweenss}{	 Sum of  squared distances between each cluster mean and sample mean. This statistic is maximized given the number of clusters.}

\item{X_name}{A character string. The actual name of the \code{X} argument.}
}
\description{
Find a frame of given size, among all possible such frames
on the input data, to minimize the minimum within-cluster
sum of square distances.
}
\details{
The method option `linear.polylog` performs
fast optimal framed clustering. The method option
 `Ckmeans.1d.dp` performs optimal framed clustering by repeatedly
 finding the best clustering within each frame.
 The `kmeans` option uses heuristic k-means algorithm in all
 frames without guarantee of clustering optimality.
}
\examples{
N <- 100
X <- rnorm(N)
K <- 5
frame.width <- 60
first.frame <- 1
last.frame <- N - frame.width + 1
method <- "linear.polylog"

result <- FramedClust( X, K, frame.width, first.frame, last.frame, method)
plot(result)

}
