% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lin_polylog_framed_clust}
\alias{lin_polylog_framed_clust}
\title{Fast and Optimal Framed Clustering}
\usage{
lin_polylog_framed_clust(X, width, K, First, Last, Prev, Next)
}
\arguments{
\item{X}{a vector of data points to be clustered}

\item{width}{the number of points in each frame}

\item{K}{the number of clusters}

\item{First}{the index to the start point in the first frame}

\item{Last}{the index to the start point of the last frame}

\item{Prev}{the index to the start point of an already clustered frame to the left of First}

\item{Next}{the index to the start point of an already clustered frame to the right of Last}
}
\value{
A dataframe containing important statistics associated with best frame

\item{ID}{Starting index of the frame with minimum SSQ}

\item{Border}{ The cluster border of K clusters}

\item{cluster}{ A vector of clusters assigned to each element in x. Each cluster is indexed by an integer from 1 to k.}

\item{centers}{ A numeric vector of the  means for each cluster.}

\item{withinss}{	A numeric vector of the  within-cluster sum of squares for each cluster.}

\item{size}{	A vector of the  number of elements in each cluster.}

\item{totss}{	Total sum of  squared distances between each element and the sample mean. This statistic is not dependent on the clustering result.}

\item{tot.withinss}{	 Total sum of  within-cluster squared distances between each element and its cluster mean. This statistic is minimized given the number of clusters.}

\item{betweenss}{	 Sum of  squared distances between each cluster mean and sample mean. This statistic is maximized given the number of clusters.}
}
\description{
The data given by vector \code{X} are clustered frame-by-frame.
Within each frame, the points are partitioned into k clusters
such that the sum of squares from points to the
assigned cluster centres is minimized.
At the minimum, all cluster centres are at the
mean of their Voronoi sets (the set of data points
which are nearest to the cluster centre).
}
\details{
For ease of programmatic exploration, k=1 is allowed, notably returning the center and withinss.

Perform framed data clustering with guaranteed optimality
  in linear-polylogarthmic time. The run time of the algorithm to find K optimal clusters in a circular data having N points is O(K N (log N)^2 ).
}
\examples{
X <- cos((-10:10))

width <- 21

K <- 2

First <- 0

Last <- 0

Prev <- -1

Next <- -1

lin_polylog_framed_clust(X, width, K, First, Last, Prev, Next)

}
