% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-FramedClust.R
\name{plot.FramedClust}
\alias{plot.FramedClust}
\title{Plot Framed Data Clustering Results}
\usage{
\method{plot}{FramedClust}(
  x,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  sub = NULL,
  col.clusters = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class \code{FramedClust} as returned by \code{FramedClust}}

\item{xlab}{a character string. The x-axis label for the plot.}

\item{ylab}{a character string. The y-axis label for the plot.}

\item{main}{a character string. The title for the plot.}

\item{sub}{a character string. The subtitle for the plot.}

\item{col.clusters}{a vector of colors, defined either by integers or by color names. If the length is shorter than the number of clusters, the colors will be reused.}

\item{...}{other arguments associated with the plot function}
}
\value{
An object of class \code{"FramedClust"},
identical to the input \code{x}
}
\description{
Visualize results of framed clustering
}
\examples{
N <- 100
X <- rchisq(N, 5)
K <- 3
frame.width <- 40

result <- FramedClust(X, K, frame.width)

plot(result)
}
