% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-FramedClust.R
\name{plot.FramedClust}
\alias{plot.FramedClust}
\title{Plot Framed Clustering Results}
\usage{
\method{plot}{FramedClust}(
  x,
  xlab = NULL,
  ylab = NULL,
  main = NULL,
  sub = NULL,
  col.clusters = NULL,
  ...
)
}
\arguments{
\item{x}{an object of class as returned by \code{CirClust}}

\item{xlab}{a character string. The x-axis label for the plot.}

\item{ylab}{a character string. The x-axis label for the plot.}

\item{main}{a character string. The title for the plot.}

\item{sub}{a character string. The subtitle for the plot.}

\item{col.clusters}{a vector of colors, defined either by integers or by color names. If the length is shorter than the number of clusters, the colors will be reused.}

\item{...}{other arguments associated with the plot function}
}
\value{
the same input object of class \code{CirClust}
}
\description{
Visualize clusters of circular data from clustering result of class \code{CirClust}.
}
\examples{

n <- 100

m <- 5

O <- c(rnorm(n,mean=5,sd=m),rnorm(n,mean=15,sd=m),rnorm(n,mean=26,sd=m))

K <- 3

Circumference <- 28

result <- CirClust(O, K, Circumference, method = "FOCC")

color <- c("#0000CD","#808080", "#DC143C")

plot(result,col.clusters = color)


}
